🐛 Fix
- Corrections des colonnes SQL

+++ back-end/ChaTop/src/main/java/com/sebastiend/ChaTop/models/entities/UserEntity.java
@@ -13,7 +13,7 @@ public class UserEntity {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(columnDefinition = "BIGINT")
+    @Column(columnDefinition = "Int")
     private Integer id;
 
     @Column(name="name")
